+++ Assets/Scripts/Tooltip/TooltipModel.cs
@@ -13,9 +13,9 @@ public readonly struct TooltipModel
     public readonly string Body;
     public readonly Sprite Icon;
     public readonly TooltipKind Kind;
-    public readonly float scoreMultiplier;
+    public readonly double scoreMultiplier;
 
-    public TooltipModel(string title, string body, Sprite icon, TooltipKind kind, float scoreMultiplier)
+    public TooltipModel(string title, string body, Sprite icon, TooltipKind kind, double scoreMultiplier)
     {
         Title = title;
         Body = body;
